refactor: move the sprite schemas into spriteValidation.js, via a shared leaf module (#3873)
The Sprite Manager block was the last large domain still inline in
validation.js. It could not be lifted as-is: `recordRenderPinFields`,
`cloudModelIdString`, `grokVideoDurationSchema` and `isSafeSubdirFilter`
are defined in validation.js's own body and consumed by six non-sprite
callers, and a domain file cannot import them back — validation.js's
`export * from` re-exports hoist above its body, so the read TDZs.

The four fragments now live in a new leaf `server/lib/sharedSchemas.js`
(the precedent zodCompat.js set for `emptyToUndefined`), with
validation.js re-exporting the same objects so existing deep imports are
untouched. The six non-sprite call sites import from the leaf directly.
