+++ .changelog/NEXT.md
@@ -47,3 +47,4 @@
 - The Pipeline's Reverse Outline map now keeps the selected scene in the URL (`?scene=<id>`) instead of component state, so the open scene detail survives a reload and can be linked to or shared. A stale id — after regenerating the outline, say — falls back to the "click a marker" placeholder rather than a blank card, and regenerating clears the selection outright.
 - The six "file too large" toasts — album covers, artist portraits, tracks, author headshots, reference audio and OpenClaw attachments — now render their size cap through the shared `formatBytes` helper instead of each dividing by 1024 twice and gluing on "MB". The rendered text gains a space before the unit (`12 MB` rather than `12MB`), which is what the rest of PortOS already shows and what the reference-audio toast already read. The OpenClaw combined-attachment message picked up the same treatment for its decoded-size estimate, so the two limits it can report no longer print their units two different ways.
 - The post-restart agent-recovery path is now its own function instead of a 65-line branch inlined at the top of `handleAgentCompletion`, so a change to either it or the live in-memory completion path no longer means reading past the other. The pause guard that path sits behind — the one keeping a stray completion event from cleaning the worktree out from under a later resume — is now pinned by tests that call the real function rather than by a source-text assertion that broke the moment the code moved.
+- The ~460-line Sprite Manager block in `server/lib/validation.js` now lives in its own `server/lib/spriteValidation.js`, matching the eleven other per-domain validation files. A straight lift wasn't possible: the block leans on four fragments (`recordRenderPinFields`, `cloudModelIdString`, `grokVideoDurationSchema`, `isSafeSubdirFilter`) that music video, universe builder, series, video gen and backup also use, and a domain file can't read them back out of `validation.js` — ESM hoists its `export * from` re-exports, so the read lands in the temporal dead zone. Those four moved to a new leaf `server/lib/sharedSchemas.js` first, with `validation.js` still re-exporting them, so every existing deep import keeps working unchanged.
