+++ server/lib/index.js
@@ -39,9 +39,14 @@ export * as postValidation from './postValidation.js';
 export * as privacyValidation from './privacyValidation.js';
 export * as roundsValidation from './roundsValidation.js';
 export * as socketValidation from './socketValidation.js';
+export * as spriteValidation from './spriteValidation.js';
 export * as storyBuilderValidation from './storyBuilderValidation.js';
 export * as telegramValidation from './telegramValidation.js';
 export * from './validation.js';
+// Cross-domain Zod fragments both `validation.js` and the per-domain files
+// import (leaf module, no cycle). Flat because `validation.js` re-exports the
+// same objects — same identity, so the collision detector is satisfied.
+export * from './sharedSchemas.js';
 export * from './threejsModel.js';
 export * from './threejsModelCoverage.js';
 export * from './threejsModelFamilies.js';
