+++ server/routes/universeBuilder/render.js
@@ -8,7 +8,8 @@
 import { Router } from 'express';
 import { z } from 'zod';
 import { asyncHandler } from '../../lib/errorHandler.js';
-import { validateRequest, cloudModelIdString } from '../../lib/validation.js';
+import { validateRequest } from '../../lib/validation.js';
+import { cloudModelIdString } from '../../lib/sharedSchemas.js';
 import { RECORD_RENDER_MODEL_MAX } from '../../lib/renderTargets.js';
 import { BIBLE_LIMITS } from '../../lib/storyBible.js';
 import * as svc from '../../services/universeBuilder.js';
